# Reduce allocations for iterators