This addresses #94 
The change is simple (and simplistic, some improvements can be done down the line): instead of allocating a new container iterator whenever we change a key, we keep a slot for one of each kind in the outer iterator structure, and we update it and point to it when we switch keys. There's a small constant memory overhead due to this embedding, but overall it should be better for memory pressure than creating and discarding many tiny objects for most bitmaps.