@lemire 2021-12-27
@Oppen It is reasonable. Do you happen to have benchmarks to quantify the benefits?

I am nervous about trusting intuition when it comes to such questions.