@Oppen 2021-12-27
Yes, but I'm running more because the confidence interval was rather weak.
For reference, I'm running the following command right now, both on the master branch and this branch:
```shell
for i in {1..100}; do go test -bench='Iter|Advance' -run -; done > bench_iter.new
```
I'll run `benchstat` on the results afterwards and post again.