@Oppen 2022-04-09
I rebased against current HEAD and run the benchmarks again 30 times each. I ran the new version first, master later. The results changed as follows:
```
name                                                                 old time/op    new time/op    delta                     
ShortIteratorAdvanceArray/init_size_1_shortIterator_advance-8          57.1ns ±16%    56.5ns ±17%     ~     (p=0.322 n=29+29)
ShortIteratorAdvanceArray/init_size_650_shortIterator_advance-8        87.0ns ±14%    88.2ns ±12%     ~     (p=0.323 n=30+28)
ShortIteratorAdvanceArray/init_size_6500_shortIterator_advance-8       98.9ns ±16%    99.0ns ±10%     ~     (p=0.638 n=29+28)
ShortIteratorAdvanceArray/init_size_65535_shortIterator_advance-8       105ns ±15%     104ns ± 6%     ~     (p=0.924 n=29+25)
ShortIteratorNextArray/init_size_1_shortIterator_next-8                46.7ns ±18%    47.4ns ±16%     ~     (p=0.495 n=30+29)
ShortIteratorNextArray/init_size_650_shortIterator_next-8              1.99µs ±19%    2.20µs ±17%  +10.59%  (p=0.000 n=29+29)
ShortIteratorNextArray/init_size_6500_shortIterator_next-8             19.8µs ±18%    22.1µs ±17%  +11.44%  (p=0.000 n=29+30)
ShortIteratorNextArray/init_size_65535_shortIterator_next-8             203µs ±20%     219µs ±10%   +8.26%  (p=0.000 n=29+28)
IterateRoaring/iterator-compressed-8                                    425µs ±20%     445µs ± 5%     ~     (p=0.082 n=28+25)
IterateRoaring/iterator-8                                               435µs ±26%     437µs ± 7%     ~     (p=0.535 n=29+25)
IterateRoaring/iterate-compressed-8                                     259µs ±20%     274µs ± 4%   +5.85%  (p=0.007 n=29+26)
IterateRoaring/iterate-8                                                285µs ±26%     315µs ±12%  +10.84%  (p=0.000 n=28+29)
SparseIterateRoaring-8                                                  498µs ±27%     458µs ±11%   -7.99%  (p=0.003 n=29+29)
SparseAdvanceRoaring/advance_from_1-8                                   110ns ±19%      85ns ± 7%  -23.05%  (p=0.000 n=28+28)
SparseAdvanceRoaring/advance_from_2-8                                   112ns ±27%      89ns ±24%  -20.92%  (p=0.000 n=28+29)
SparseAdvanceRoaring/advance_from_65-8                                  109ns ±15%      92ns ±23%  -16.26%  (p=0.000 n=28+30)
SparseAdvanceRoaring/advance_from_650-8                                 108ns ±19%      92ns ±24%  -15.01%  (p=0.000 n=27+30)
SparseAdvanceOnHugeData-8                                              1.04µs ±16%    0.68µs ±12%  -34.80%  (p=0.000 n=28+27)
SparseAdvanceSequentially/advance_from_1-8                              119µs ± 6%     120µs ± 4%     ~     (p=0.188 n=29+29)
SparseAdvanceSequentially/advance_from_2-8                              113µs ±10%     114µs ± 3%     ~     (p=0.940 n=29+26)
SparseAdvanceSequentially/advance_from_65-8                             345µs ±18%     373µs ± 7%   +8.28%  (p=0.000 n=29+28)
SparseAdvanceSequentially/advance_from_650-8                            342µs ±28%     366µs ± 5%   +6.86%  (p=0.002 n=30+26)
IterateBitset-8                                                         200µs ±24%     198µs ± 5%     ~     (p=0.527 n=29+25)
SparseIterateBitset-8                                                  1.93ms ±14%    1.87ms ± 4%   -3.09%  (p=0.011 n=28+25)
ShortIteratorAdvanceBitmap/init_size_1_shortIterator_advance-8         42.7ns ±17%    42.2ns ± 9%     ~     (p=0.666 n=30+27)
ShortIteratorAdvanceBitmap/init_size_650_shortIterator_advance-8       45.4ns ±16%    45.0ns ±14%     ~     (p=0.455 n=29+29)
ShortIteratorAdvanceBitmap/init_size_6500_shortIterator_advance-8      45.2ns ±18%    44.6ns ±13%     ~     (p=0.516 n=30+26)
ShortIteratorAdvanceBitmap/init_size_65535_shortIterator_advance-8     45.2ns ±20%    44.4ns ± 8%     ~     (p=0.240 n=30+27)
ShortIteratorNextBitmap/init_size_1_shortIterator_next-8                629ns ±15%     419ns ± 4%  -33.31%  (p=0.000 n=29+24)
ShortIteratorNextBitmap/init_size_650_shortIterator_next-8             3.98µs ±17%    3.76µs ±10%   -5.46%  (p=0.001 n=30+27)
ShortIteratorNextBitmap/init_size_6500_shortIterator_next-8            34.1µs ±13%    34.2µs ±11%     ~     (p=0.464 n=29+27)
ShortIteratorNextBitmap/init_size_65535_shortIterator_next-8            336µs ±10%     337µs ± 8%     ~     (p=0.457 n=24+25)
ShortIteratorAdvanceRuntime/init_size_1_shortIterator_advance-8        46.8ns ±15%    46.4ns ±14%     ~     (p=0.514 n=29+29)
ShortIteratorAdvanceRuntime/init_size_650_shortIterator_advance-8      52.6ns ±23%    52.3ns ±14%     ~     (p=0.918 n=30+30)
ShortIteratorAdvanceRuntime/init_size_6500_shortIterator_advance-8     52.4ns ±16%    52.7ns ±10%     ~     (p=0.363 n=29+30)
ShortIteratorAdvanceRuntime/init_size_65535_shortIterator_advance-8    52.3ns ±12%    53.2ns ±12%     ~     (p=0.149 n=30+29)
ShortIteratorNextRuntime/init_size_1_shortIterator_next-8              42.9ns ±14%    42.8ns ±12%     ~     (p=0.844 n=29+30)
ShortIteratorNextRuntime/init_size_650_shortIterator_next-8            3.23µs ±15%    3.25µs ±10%     ~     (p=0.655 n=29+26)
ShortIteratorNextRuntime/init_size_6500_shortIterator_next-8           32.1µs ±24%    31.7µs ± 7%     ~     (p=0.445 n=29+25)
ShortIteratorNextRuntime/init_size_65535_shortIterator_next-8           325µs ±22%     320µs ± 8%     ~     (p=0.279 n=29+26)

name                                                                 old alloc/op   new alloc/op   delta
IterateRoaring/iterator-compressed-8                                    96.0B ± 0%    112.0B ± 0%  +16.67%  (p=0.000 n=30+30)
IterateRoaring/iterator-8                                               96.0B ± 0%    112.0B ± 0%  +16.67%  (p=0.000 n=30+30)
IterateRoaring/iterate-compressed-8                                     0.00B          0.00B          ~     (all equal)
IterateRoaring/iterate-8                                                0.00B          0.00B          ~     (all equal)
SparseAdvanceRoaring/advance_from_1-8                                   64.0B ± 0%    112.0B ± 0%  +75.00%  (p=0.000 n=30+30)
SparseAdvanceRoaring/advance_from_2-8                                   64.0B ± 0%    112.0B ± 0%  +75.00%  (p=0.000 n=30+30)
SparseAdvanceRoaring/advance_from_65-8                                  64.0B ± 0%    112.0B ± 0%  +75.00%  (p=0.000 n=30+30)
SparseAdvanceRoaring/advance_from_650-8                                 64.0B ± 0%    112.0B ± 0%  +75.00%  (p=0.000 n=30+30)
SparseAdvanceOnHugeData-8                                                549B ±20%      112B ± 0%  -79.59%  (p=0.000 n=28+30)
SparseAdvanceSequentially/advance_from_1-8                              64.0B ± 0%    112.0B ± 0%  +75.00%  (p=0.000 n=30+30)
SparseAdvanceSequentially/advance_from_2-8                              64.0B ± 0%    112.0B ± 0%  +75.00%  (p=0.000 n=30+30)
SparseAdvanceSequentially/advance_from_65-8                             64.0B ± 0%    112.0B ± 0%  +75.00%  (p=0.000 n=30+30)
SparseAdvanceSequentially/advance_from_650-8                            64.0B ± 0%    112.0B ± 0%  +75.00%  (p=0.000 n=30+30)

name                                                                 old allocs/op  new allocs/op  delta
IterateRoaring/iterator-compressed-8                                     4.00 ± 0%      1.00 ± 0%  -75.00%  (p=0.000 n=30+30)
IterateRoaring/iterator-8                                                4.00 ± 0%      1.00 ± 0%  -75.00%  (p=0.000 n=30+30)
IterateRoaring/iterate-compressed-8                                      0.00           0.00          ~     (all equal)
IterateRoaring/iterate-8                                                 0.00           0.00          ~     (all equal)
SparseAdvanceRoaring/advance_from_1-8                                    2.00 ± 0%      1.00 ± 0%  -50.00%  (p=0.000 n=30+30)
SparseAdvanceRoaring/advance_from_2-8                                    2.00 ± 0%      1.00 ± 0%  -50.00%  (p=0.000 n=30+30)
SparseAdvanceRoaring/advance_from_65-8                                   2.00 ± 0%      1.00 ± 0%  -50.00%  (p=0.000 n=30+30)
SparseAdvanceRoaring/advance_from_650-8                                  2.00 ± 0%      1.00 ± 0%  -50.00%  (p=0.000 n=30+30)
SparseAdvanceOnHugeData-8                                                23.8 ±24%       1.0 ± 0%  -95.79%  (p=0.000 n=30+30)
SparseAdvanceSequentially/advance_from_1-8                               2.00 ± 0%      1.00 ± 0%  -50.00%  (p=0.000 n=30+30)
SparseAdvanceSequentially/advance_from_2-8                               2.00 ± 0%      1.00 ± 0%  -50.00%  (p=0.000 n=30+30)
SparseAdvanceSequentially/advance_from_65-8                              2.00 ± 0%      1.00 ± 0%  -50.00%  (p=0.000 n=30+30)
SparseAdvanceSequentially/advance_from_650-8                             2.00 ± 0%      1.00 ± 0%  -50.00%  (p=0.000 n=30+30)
```

Specifically, the one test that took 40% longer before takes 30% less time now, so I can only guess, given the order of magnitude of time (nsecs) that the difference is due to throttling or scheduling effects.
I feel more comfortable merging now.

I'd like to make some improvements to tend to zero-alloc or at least make allocs smaller (for example, by keeping only one index into the container instead of one per iterator), but that can be done in a future PR.